v0.95 - Major restructuring
- Major restructure of the JS-AHK communication. It is now much more
direct (and presumably faster) since it uses the new features of
**AutoHotkey v1.1.17.01**.
- Major restructure of the HTML host. It is now a virtual page that
works in IE8+. It should be free of race conditions and window.onload
works fine.
- Major code reshuffle. Removed all libraries except "API.ahk" from
`lib`. Reduced a lot of code boilerplate.
- Implemented all g-labels. They now have a similar behavior as in AHK:
you only need to declare `function GuiClose(){}` and it will be
auto-activated when the time comes.
- `GuiClose` and `OnExit` can prevent ExitApp by returning false (or 0)
- Functions that have no return value (therefore `undefined`) now return
the empty string. This was done for the sake of code brevity and it
should be harmless.
- Renamed `Parameters` to `A_Args`
- `OnMessage` now follows the documentation
- Marked `SetTimer` as "Cannot use yet"
- Added a JS specific function: `Require()`
- Fixed a lot of bugs in API
- Added Lexikos in the About section